Cleanup SellCancellation
