+++ .Lib9c.Tests/Action/SellCancellation5Test.cs
@@ -1,7 +1,8 @@
 namespace Lib9c.Tests.Action
 {
     using System;
-    using System.Linq;
+    using System.Collections.Generic;
+    using Bencodex.Types;
     using Libplanet;
     using Libplanet.Action;
     using Libplanet.Assets;
@@ -10,6 +11,7 @@ namespace Lib9c.Tests.Action
     using Nekoyume.Action;
     using Nekoyume.Model;
     using Nekoyume.Model.Item;
+    using Nekoyume.Model.Mail;
     using Nekoyume.Model.State;
     using Serilog;
     using Xunit;
@@ -20,6 +22,8 @@ namespace Lib9c.Tests.Action
         private readonly IAccountStateDelta _initialState;
         private readonly Address _agentAddress;
         private readonly Address _avatarAddress;
+        private readonly GoldCurrencyState _goldCurrencyState;
+        private readonly TableSheets _tableSheets;
 
         public SellCancellation5Test(ITestOutputHelper outputHelper)
         {
@@ -36,10 +40,10 @@ namespace Lib9c.Tests.Action
                     .SetState(Addresses.TableSheet.Derive(key), value.Serialize());
             }
 
-            var tableSheets = new TableSheets(sheets);
+            _tableSheets = new TableSheets(sheets);
 
             var currency = new Currency("NCG", 2, minters: null);
-            var goldCurrencyState = new GoldCurrencyState(currency);
+            _goldCurrencyState = new GoldCurrencyState(currency);
 
             _agentAddress = new PrivateKey().ToAddress();
             var agentState = new AgentState(_agentAddress);
@@ -49,71 +53,291 @@ namespace Lib9c.Tests.Action
                 _avatarAddress,
                 _agentAddress,
                 0,
-                tableSheets.GetAvatarSheets(),
+                _tableSheets.GetAvatarSheets(),
                 new GameConfigState(),
                 rankingMapAddress)
             {
                 worldInformation = new WorldInformation(
                     0,
-                    tableSheets.WorldSheet,
+                    _tableSheets.WorldSheet,
                     GameConfig.RequireClearedStageLevel.ActionsInShop),
             };
             agentState.avatarAddresses[0] = _avatarAddress;
 
-            var equipment = ItemFactory.CreateItemUsable(
-                tableSheets.EquipmentItemSheet.First,
-                Guid.NewGuid(),
-                0);
-            var shopState = new ShopState();
-            shopState.Register(new ShopItem(
-                _agentAddress,
-                _avatarAddress,
-                Guid.NewGuid(),
-                new FungibleAssetValue(goldCurrencyState.Currency, 100, 0),
-                equipment));
-
             _initialState = _initialState
-                .SetState(GoldCurrencyState.Address, goldCurrencyState.Serialize())
-                .SetState(Addresses.Shop, shopState.Serialize())
+                .SetState(GoldCurrencyState.Address, _goldCurrencyState.Serialize())
                 .SetState(_agentAddress, agentState.Serialize())
+                .SetState(Addresses.Shop, new ShopState().Serialize())
                 .SetState(_avatarAddress, avatarState.Serialize());
         }
 
-        [Fact]
-        public void Execute()
+        [Theory]
+        [InlineData(ItemType.Equipment, "F9168C5E-CEB2-4faa-B6BF-329BF39FA1E4", true)]
+        [InlineData(ItemType.Costume, "936DA01F-9ABD-4d9d-80C7-02AF85C822A8", true)]
+        [InlineData(ItemType.Equipment, "F9168C5E-CEB2-4faa-B6BF-329BF39FA1E4", false)]
+        [InlineData(ItemType.Costume, "936DA01F-9ABD-4d9d-80C7-02AF85C822A8", false)]
+        public void Execute(ItemType itemType, string guid, bool contain)
         {
-            var shopState = _initialState.GetShopState();
-            Assert.Single(shopState.Products);
-
-            var (_, shopItem) = shopState.Products.FirstOrDefault();
-            Assert.NotNull(shopItem);
-
             var avatarState = _initialState.GetAvatarState(_avatarAddress);
-            Assert.False(avatarState.inventory.TryGetNonFungibleItem(
-                shopItem.ItemUsable.ItemId,
-                out ItemUsable _));
+            INonFungibleItem nonFungibleItem;
+            Guid itemId = new Guid(guid);
+            Guid productId = itemId;
+            ItemSubType itemSubType;
+            const long requiredBlockIndex = 0;
+            ShopState legacyShopState = _initialState.GetShopState();
+            if (itemType == ItemType.Equipment)
+            {
+                var itemUsable = ItemFactory.CreateItemUsable(
+                    _tableSheets.EquipmentItemSheet.First,
+                    itemId,
+                    requiredBlockIndex);
+                nonFungibleItem = itemUsable;
+                itemSubType = itemUsable.ItemSubType;
+            }
+            else
+            {
+                var costume = ItemFactory.CreateCostume(_tableSheets.CostumeItemSheet.First, itemId);
+                costume.Update(requiredBlockIndex);
+                nonFungibleItem = costume;
+                itemSubType = costume.ItemSubType;
+            }
+
+            var result = new DailyReward.DailyRewardResult()
+            {
+                id = default,
+                materials = new Dictionary<Material, int>(),
+            };
+
+            for (var i = 0; i < 100; i++)
+            {
+                var mail = new DailyRewardMail(result, i, default, 0);
+                avatarState.Update(mail);
+            }
+
+            Address shardedShopAddress = ShardedShopState.DeriveAddress(itemSubType, productId);
+            ShardedShopState shopState = new ShardedShopState(shardedShopAddress);
+            var shopItem = new ShopItem(
+                _agentAddress,
+                _avatarAddress,
+                productId,
+                new FungibleAssetValue(_goldCurrencyState.Currency, 100, 0),
+                requiredBlockIndex,
+                nonFungibleItem);
+
+            if (contain)
+            {
+                shopState.Register(shopItem);
+                avatarState.inventory.AddItem((ItemBase)nonFungibleItem);
+                Assert.Empty(legacyShopState.Products);
+                Assert.Single(shopState.Products);
+            }
+            else
+            {
+                legacyShopState.Register(shopItem);
+                Assert.Single(legacyShopState.Products);
+                Assert.Empty(shopState.Products);
+            }
+
+            Assert.Equal(requiredBlockIndex, nonFungibleItem.RequiredBlockIndex);
+            Assert.Equal(contain, avatarState.inventory.TryGetNonFungibleItem(itemId, out _));
+
+            IAccountStateDelta prevState = _initialState
+                .SetState(_avatarAddress, avatarState.Serialize())
+                .SetState(Addresses.Shop, legacyShopState.Serialize())
+                .SetState(shardedShopAddress, shopState.Serialize());
 
             var sellCancellationAction = new SellCancellation5
             {
                 productId = shopItem.ProductId,
                 sellerAvatarAddress = _avatarAddress,
+                itemSubType = itemSubType,
             };
             var nextState = sellCancellationAction.Execute(new ActionContext
             {
-                BlockIndex = 0,
-                PreviousStates = _initialState,
+                BlockIndex = 1,
+                PreviousStates = prevState,
                 Random = new TestRandom(),
                 Rehearsal = false,
                 Signer = _agentAddress,
             });
 
-            var nextShopState = nextState.GetShopState();
+            ShardedShopState nextShopState = new ShardedShopState((Dictionary)nextState.GetState(shardedShopAddress));
             Assert.Empty(nextShopState.Products);
 
             var nextAvatarState = nextState.GetAvatarState(_avatarAddress);
-            Assert.True(nextAvatarState.inventory.TryGetNonFungibleItem(
-                shopItem.ItemUsable.ItemId,
-                out ItemUsable _));
+            Assert.True(nextAvatarState.inventory.TryGetNonFungibleItem(itemId, out INonFungibleItem nextNonFungibleItem));
+            Assert.Equal(1, nextNonFungibleItem.RequiredBlockIndex);
+            Assert.Equal(30, nextAvatarState.mailBox.Count);
+            ShopState nextLegacyShopState = nextState.GetShopState();
+            Assert.Empty(nextLegacyShopState.Products);
+        }
+
+        [Fact]
+        public void Execute_Throw_FailedLoadStateException()
+        {
+            var action = new SellCancellation5
+            {
+                productId = default,
+                sellerAvatarAddress = default,
+                itemSubType = ItemSubType.Weapon,
+            };
+
+            Assert.Throws<FailedLoadStateException>(() => action.Execute(new ActionContext()
+                {
+                    BlockIndex = 0,
+                    PreviousStates = _initialState,
+                    Random = new TestRandom(),
+                    Signer = default,
+                })
+            );
+        }
+
+        [Fact]
+        public void Execute_Throw_NotEnoughClearedStageLevelException()
+        {
+            var avatarState = new AvatarState(_initialState.GetAvatarState(_avatarAddress))
+            {
+                worldInformation = new WorldInformation(
+                    0,
+                    _tableSheets.WorldSheet,
+                    0
+                ),
+            };
+
+            IAccountStateDelta prevState = _initialState.SetState(_avatarAddress, avatarState.Serialize());
+
+            var action = new SellCancellation5
+            {
+                productId = default,
+                sellerAvatarAddress = _avatarAddress,
+                itemSubType = ItemSubType.Weapon,
+            };
+
+            Assert.Throws<NotEnoughClearedStageLevelException>(() => action.Execute(new ActionContext
+            {
+                BlockIndex = 0,
+                PreviousStates = prevState,
+                Signer = _agentAddress,
+            }));
+        }
+
+        [Fact]
+        public void Execute_Throw_ItemDoesNotExistException()
+        {
+            Guid productId = Guid.NewGuid();
+            ItemUsable itemUsable = ItemFactory.CreateItemUsable(
+                _tableSheets.EquipmentItemSheet.First,
+                Guid.NewGuid(),
+                Sell.ExpiredBlockIndex);
+            Address shardedShopAddress = ShardedShopState.DeriveAddress(itemUsable.ItemSubType, productId);
+            ShardedShopState shopState = new ShardedShopState(shardedShopAddress);
+
+            var shopItem = new ShopItem(
+                _agentAddress,
+                _avatarAddress,
+                productId,
+                new FungibleAssetValue(_goldCurrencyState.Currency, 100, 0),
+                Sell.ExpiredBlockIndex,
+                itemUsable);
+
+            IAccountStateDelta prevState = _initialState
+                .SetState(shardedShopAddress, shopState.Serialize());
+
+            var action = new SellCancellation5
+            {
+                productId = shopItem.ProductId,
+                sellerAvatarAddress = _avatarAddress,
+                itemSubType = itemUsable.ItemSubType,
+            };
+
+            Assert.Throws<ItemDoesNotExistException>(() => action.Execute(new ActionContext()
+                {
+                    BlockIndex = 0,
+                    PreviousStates = prevState,
+                    Random = new TestRandom(),
+                    Signer = _agentAddress,
+                })
+            );
+        }
+
+        [Fact]
+        public void Execute_Throw_InvalidAddressException_From_Agent()
+        {
+            Guid productId = Guid.NewGuid();
+            ItemUsable itemUsable = ItemFactory.CreateItemUsable(
+                _tableSheets.EquipmentItemSheet.First,
+                Guid.NewGuid(),
+                Sell.ExpiredBlockIndex);
+            Address shardedShopAddress = ShardedShopState.DeriveAddress(itemUsable.ItemSubType, productId);
+            ShardedShopState shopState = new ShardedShopState(shardedShopAddress);
+
+            var shopItem = new ShopItem(
+                default,
+                _avatarAddress,
+                productId,
+                new FungibleAssetValue(_goldCurrencyState.Currency, 100, 0),
+                Sell.ExpiredBlockIndex,
+                itemUsable);
+            shopState.Register(shopItem);
+
+            IAccountStateDelta prevState = _initialState
+                .SetState(shardedShopAddress, shopState.Serialize());
+
+            var action = new SellCancellation5
+            {
+                productId = shopItem.ProductId,
+                sellerAvatarAddress = _avatarAddress,
+                itemSubType = itemUsable.ItemSubType,
+            };
+
+            Assert.Throws<InvalidAddressException>(() => action.Execute(new ActionContext()
+                {
+                    BlockIndex = 0,
+                    PreviousStates = prevState,
+                    Random = new TestRandom(),
+                    Signer = _agentAddress,
+                })
+            );
+        }
+
+        [Fact]
+        public void Execute_Throw_InvalidAddressException_From_Avatar()
+        {
+            Guid productId = Guid.NewGuid();
+            ItemUsable itemUsable = ItemFactory.CreateItemUsable(
+                _tableSheets.EquipmentItemSheet.First,
+                Guid.NewGuid(),
+                Sell.ExpiredBlockIndex);
+            Address shardedShopAddress = ShardedShopState.DeriveAddress(itemUsable.ItemSubType, productId);
+            ShardedShopState shopState = new ShardedShopState(shardedShopAddress);
+
+            var shopItem = new ShopItem(
+                _agentAddress,
+                default,
+                productId,
+                new FungibleAssetValue(_goldCurrencyState.Currency, 100, 0),
+                Sell.ExpiredBlockIndex,
+                itemUsable);
+            shopState.Register(shopItem);
+
+            IAccountStateDelta prevState = _initialState
+                .SetState(shardedShopAddress, shopState.Serialize());
+
+            var action = new SellCancellation5
+            {
+                productId = shopItem.ProductId,
+                sellerAvatarAddress = _avatarAddress,
+                itemSubType = itemUsable.ItemSubType,
+            };
+
+            Assert.Throws<InvalidAddressException>(() => action.Execute(new ActionContext()
+                {
+                    BlockIndex = 0,
+                    PreviousStates = prevState,
+                    Random = new TestRandom(),
+                    Signer = _agentAddress,
+                })
+            );
         }
     }
 }
